// InvoiceWebApp-Material/src/app/shared/models/settings.model.ts
export default class Settings {
    id: number = null;

    company_name: string = null;
    website: string = null;
    phone: string = null;

    address: string = null;
    postal_code: string = null;
    city: string = null;
    country: string = null;

    business_number: string = null;
    vat: string = null;

    bank: string = null;
    bank_account: string = null;

    email: string = null;
    password: string = null;
    smtp: string = null;
    smtp_port: number = null;

    invoice_prefix: string = null;

    logo: string = null;
    show_logo: boolean = null;
    show_logo_in_pdf: boolean = null;
}